+++ Editor/Shared/Assets/ObservableAssets.Gen.cs
@@ -12,11 +12,13 @@ namespace Unity.Services.Deployment.Editor.Shared.Assets
     {
         readonly AssetPostprocessorProxy m_Postprocessor;
         protected readonly Dictionary<string, T> m_AssetPaths = new Dictionary<string, T>();
+        readonly IReadOnlyList<string> m_Extensions;
 
-        public ObservableAssets() : this(new AssetPostprocessorProxy(), true) {}
+        public ObservableAssets(IReadOnlyList<string> extensions) : this(extensions, new AssetPostprocessorProxy(), true) {}
 
-        public ObservableAssets(AssetPostprocessorProxy assetPostprocessor, bool loadAssets)
+        internal ObservableAssets(IReadOnlyList<string> extensions, AssetPostprocessorProxy assetPostprocessor, bool loadAssets)
         {
+            m_Extensions = extensions;
             m_Postprocessor = assetPostprocessor;
             m_Postprocessor.AllAssetsPostprocessed += AllAssetsPostprocessed;
             if (loadAssets)
@@ -63,6 +65,9 @@ namespace Unity.Services.Deployment.Editor.Shared.Assets
 
             foreach (var imported in args.ImportedAssetPaths)
             {
+                if (!ShouldProcess(imported))
+                    continue;
+
                 var asset = AssetDatabase.LoadAssetAtPath<T>(imported);
                 if (asset != null)
                 {
@@ -88,6 +93,25 @@ namespace Unity.Services.Deployment.Editor.Shared.Assets
                     MovePath(movedToPath, movedFromPath);
                 }
             }
+        }
+
+        bool ShouldProcess(string importedPath)
+        {
+            if (m_Extensions != null && m_Extensions.Any())
+            {
+                var matchingExt = m_Extensions.Any(ext => importedPath.EndsWith(ext));
+                if (!matchingExt)
+                    return false; // Asset does not match any given extension, skip
+            }
+            else
+            {
+                // If we dont have a limited set of extensions, use the the DB type
+                var assetType = AssetDatabase.GetMainAssetTypeAtPath(importedPath);
+                if (assetType == null || !typeof(T).IsAssignableFrom(assetType))
+                    return false;
+            }
+
+            return true;
         }
 
         protected virtual void Dispose(bool disposing)