+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to the Deployment package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.6.2] - 2025-09-10
+
+### Fixed
+- Fixed assets being loaded despite being of the incorrect type
+
+## [1.6.0] - 2025-08-06
+
+### Added
+- Added an Inspector header for Deployment assets that displays their current status.
+
 ## [1.5.0] - 2025-07-03
 
 ### Fixed